Renamed generator top
Renamed generator.top to generator.factory to make its role more explicit
